Fix comparison

+++ src/umpire/util/FixedMallocPool.cpp
@@ -48,7 +48,7 @@ FixedMallocPool::FixedMallocPool(const std::size_t object_bytes, const std::size
   // Free-list links are stored in the slots themselves
   UMPIRE_ASSERT(object_bytes >= sizeof(unsigned int));
   // Slot indices (free-list links and num_used) are unsigned int
-  UMPIRE_ASSERT(objects_per_pool <= std::numeric_limits<unsigned int>::max());
+  UMPIRE_ASSERT(objects_per_pool <= static_cast<std::size_t>(std::numeric_limits<unsigned int>::max()));
   newPool();
 }
